+++ dashboard/src/main/home/app-dashboard/new-app-flow/tabs/ReleaseTabs.tsx
@@ -8,17 +8,22 @@ import { Context } from "shared/Context";
 import Checkbox from "components/porter/Checkbox";
 import Text from "components/porter/Text";
 import { DATABASE_HEIGHT_DISABLED, DATABASE_HEIGHT_ENABLED } from "./utils";
+import InputSlider from "components/porter/InputSlider";
 
 interface Props {
     service: ReleaseService;
     editService: (service: ReleaseService) => void;
     setHeight: (height: Height) => void;
+    maxRAM: number;
+    maxCPU: number;
 }
 
 const ReleaseTabs: React.FC<Props> = ({
     service,
     editService,
     setHeight,
+    maxRAM,
+    maxCPU,
 }) => {
     const [currentTab, setCurrentTab] = React.useState<string>('main');
     const { currentCluster } = useContext(Context);
@@ -46,23 +51,33 @@ const ReleaseTabs: React.FC<Props> = ({
         return (
             <>
                 <Spacer y={1} />
-                <Input
-                    label="CPU (Millicores)"
-                    placeholder="ex: 0.5"
-                    value={service.cpu.value}
+                <InputSlider
+                    label="CPUs: "
+                    unit="Cores"
+                    min={0}
+                    max={maxCPU}
+                    color={"#3a48ca"}
+                    value={(service.cpu.value / 1000).toString()}
+                    setValue={(e) => {
+                        editService({ ...service, cpu: { readOnly: false, value: e * 1000 } });
+                    }}
+                    step={0.01}
                     disabled={service.cpu.readOnly}
-                    width="300px"
-                    setValue={(e) => { editService({ ...service, cpu: { readOnly: false, value: e } }) }}
                     disabledTooltip={"You may only edit this field in your porter.yaml."}
                 />
                 <Spacer y={1} />
-                <Input
-                    label="RAM (MB)"
-                    placeholder="ex: 1"
-                    value={service.ram.value}
+                <InputSlider
+                    label="RAM: "
+                    unit="GiB"
+                    min={0}
+                    max={maxRAM}
+                    color={"#3a48ca"}
+                    value={(service.ram.value / 1024).toString()}
+                    setValue={(e) => {
+                        editService({ ...service, ram: { readOnly: false, value: e * 1024 } });
+                    }}
                     disabled={service.ram.readOnly}
-                    width="300px"
-                    setValue={(e) => { editService({ ...service, ram: { readOnly: false, value: e } }) }}
+                    step={0.01}
                     disabledTooltip={"You may only edit this field in your porter.yaml."}
                 />
             </>
